requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
email-validator==2.1.0
python-dotenv==1.0.0
diffusers==0.21.4
huggingface_hub==0.19.4
accelerate==0.23.0
safetensors==0.4.0